% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mshap.R
\name{mshap}
\alias{mshap}
\title{mSHAP}
\usage{
mshap(shap_1, shap_2, ex_1, ex_2)
}
\arguments{
\item{shap_1}{SHAP values for the data set from the first model}

\item{shap_2}{SHAP values for the data set from the second model}

\item{ex_1}{the expected value of the explainer for the first model}

\item{ex_2}{the expected value of the explainer for the second model}
}
\value{
a list containing the multiplied SHAP values and the expected value
}
\description{
A function for calculation SHAP values of two part models
}
\details{
This function allows the user to input the SHAP values for two separate
models (along with the expected values), and mSHAP then outputs the SHAP
values of the two model predictions multiplied together.
}
